# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GraphDB Internal API
  version: 1.0.0
  description: Internal object-store backed multi-tenant graph database API.
servers:
  - url: http://localhost:8080
security:
  - TenantHeader: []
paths:
...
  /v1/commits:
    post:
      summary: Apply an atomic direct mutation commit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
      responses:
        '200':
          description: Commit applied, skipped by md5, or idempotently replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '429':
          $ref: '#/components/responses/WriteBackpressure'
  /v1/ingest/batches:
    post:
      summary: Apply a collector ingestion batch with idempotency and partial failure reporting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Batch accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '207':
          description: Batch accepted with item failures.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '429':
          $ref: '#/components/responses/WriteBackpressure'
  /v1/query:
    post:
      summary: Execute the JSON query DSL.
      parameters:
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
  /v1/query/stream:
    post:
      summary: Execute a query and stream newline-delimited JSON pages.
      parameters:
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: NDJSON stream.
          content:
            application/x-ndjson:
              schema:
                type: string
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
  /v1/query/gql:
    post:
      summary: Execute a text GQL query compiled to the JSON query DSL.
      parameters:
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GQLQueryRequest'
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: Query result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
  /v1/query/gql/stream:
    post:
      summary: Execute a text GQL query and stream newline-delimited JSON pages.
      parameters:
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GQLQueryRequest'
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: NDJSON stream.
          content:
            application/x-ndjson:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
...
  /v1/queries/running/{query_id}:
    delete:
      summary: Cancel a running query on this process.
      parameters:
        - name: query_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Query cancellation requested.
          content:
            application/json:
              schema:
                type: object
                properties:
                  killed:
                    type: boolean
                  query:
                    $ref: '#/components/schemas/RunningQuery'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/tenants:
    get:
      summary: List tenants.
      parameters:
        - name: include_legacy
          in: query
          required: false
          schema:
            type: boolean
          description: Scan legacy tenant prefixes when true; default lists only lifecycle-managed tenants.
      responses:
        '200':
          description: Tenant list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/TenantInfo'
    post:
      summary: Create or materialize tenant metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantRequest'
      responses:
        '200':
          description: Tenant created or already present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
  /v1/entities:
    get:
      summary: Page current entities by kind, source, or shard.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Shard'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: Entity page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
  /v1/entities/stream:
    get:
      summary: Stream current entity pages as newline-delimited JSON.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Shard'
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: NDJSON stream.
          content:
            application/x-ndjson:
              schema:
                type: string
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/entities/{id}:
    get:
      summary: Get one entity by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: Entity result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: integer
                  entity:
                    $ref: '#/components/schemas/Entity'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
  /v1/edges:
    get:
      summary: Page current edges by type and from shard.
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: from_shard
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: Edge page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgePage'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/edges/stream:
    get:
      summary: Stream current edge pages as newline-delimited JSON.
      parameters:
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: NDJSON stream.
          content:
            application/x-ndjson:
              schema:
                type: string
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/export/snapshot:
    get:
      summary: Export the current snapshot page.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: Snapshot export page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotPage'
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/export/snapshot/stream:
    get:
      summary: Stream the current snapshot as newline-delimited JSON.
      description: Defaults to a refs stream for sharded snapshot catalogs. Use inline=true for row-by-row entity and edge records.
      parameters:
        - name: inline
          in: query
          schema:
            type: boolean
          description: Return row-by-row entity and edge records instead of object references.
        - $ref: '#/components/parameters/ReadMinVersion'
        - $ref: '#/components/parameters/ReadAllowStale'
        - $ref: '#/components/parameters/ReadMinVersionHeader'
        - $ref: '#/components/parameters/ReadAllowStaleHeader'
      responses:
        '200':
          description: NDJSON stream.
          content:
            application/x-ndjson:
              schema:
                type: string
        '503':
          $ref: '#/components/responses/ReaderNotFresh'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/indexes:
    get:
      summary: Get the persisted index catalog.
      responses:
        '200':
          description: Index catalog.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexCatalog'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/indexes/rebuild:
    post:
      summary: Rebuild persisted indexes.
      parameters:
        - name: async
          in: query
          schema:
            type: boolean
          description: Start a background Parquet rebuild.
      responses:
        '200':
          description: Synchronous rebuild completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexCatalog'
        '202':
          description: Background rebuild task started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
  /v1/indexes/tasks/{task_id}:
    get:
      summary: Get a background index rebuild task.
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Index task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexTask'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/control/reader-fleet-readiness:
    get:
      summary: Aggregate reader heartbeat readiness for this tenant.
      parameters:
        - name: min_version
          in: query
          schema:
            type: integer
          description: Version every ready reader must be able to serve; defaults to current writer manifest version.
        - name: min_ready
          in: query
          schema:
            type: integer
          description: Minimum ready readers required; default 1.
        - name: max_staleness_ms
          in: query
          schema:
            type: integer
          description: Heartbeat freshness window; default 30000.
      responses:
        '200':
          description: Reader fleet readiness.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReaderFleetReadiness'
  /v1/control/reader-traffic-gate:
    get:
      summary: Deployment-facing local reader traffic gate.
      parameters:
        - name: min_version
          in: query
          schema:
            type: integer
          description: Version this reader must serve; defaults to current writer manifest version.
        - name: max_staleness_ms
          in: query
          schema:
            type: integer
          description: Staleness window used only when allow_stale=true; default 30000.
        - name: allow_stale
          in: query
          schema:
            type: boolean
            default: false
          description: When true, bounded stale readers may keep traffic until max_staleness_ms is exceeded.
        - name: refresh
          in: query
          schema:
            type: boolean
            default: true
          description: When true, a lagging cached reader attempts to load at least the target version before deciding.
      responses:
        '200':
          description: This reader can receive traffic for the tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReaderTrafficGate'
        '503':
          description: This reader should be drained from traffic for the tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReaderTrafficGate'
  /v1/control/recover:
    post:
      summary: Recover contiguous orphan commits into the tenant manifest.
      responses:
        '200':
          description: Recovery report.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
  parameters:
...
    Kind:
      name: kind
      in: query
      schema:
        type: string
...
    ReadMinVersion:
      name: min_version
      in: query
      schema:
        type: integer
        minimum: 0
      description: Require the reader-visible snapshot to be at least this version.
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    MethodNotAllowed:
      description: Operation disabled in this mode.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CommitRequest:
      type: object
      required: [mutations]
      properties:
        expected_version:
          type: integer
        idempotency_key:
          type: string
          description: Optional direct commit idempotency key. Same key must reuse the same request body.
        mutations:
          $ref: '#/components/schemas/Mutations'
    CommitResult:
      type: object
      properties:
        tenant_id:
          type: string
        version:
          type: integer
        readable_version:
          type: integer
        read_after_commit_id:
          type: string
        skipped:
          type: boolean
        idempotent_replay:
          type: boolean
        data_md5:
          type: string
        suppressed:
          type: array
          items:
            $ref: '#/components/schemas/Conflict'
...